Fix alignment of vertical slideshow images
Not all of the vertical images were being center aligned because I had
reset the `figure`s to `max-width: 100%` instead of `width: 100%`, so
the images were centered in their containers but the containers weren’t
centered. Also fixed odd extra padding on first image - there’s
probably a reason it was set up that way though.