Fix scoping and naming mix ups
“Name” is a reserved word so it shouldn’t be used as a var name anyway,
but it was being used both as a project ID and as a form field name
(which might change independent of project name). Change form field
name to “bioName” and project identified to “displayName”. Also changed
loop `entries` to `items` because since it used the same name for the
local and global instances the draft and published entries weren’t
ending up in a common array.